fix: 画面階層減らす。lazy-loadingからloading.tsxに変更.
