+++ src/app/[eventId]/expense/page.tsx
@@ -1,23 +1,32 @@
-'use client';
-
-import { useAtom } from 'jotai';
+import { doc, getDoc } from 'firebase/firestore';
 import { NextPage } from 'next';
-import { lazy, Suspense } from 'react';
 
-import { eventAtom } from '@/atoms/eventAtom';
+import ExpensePageContent from '@/components/containers/expense/ExpensePageContent';
+import { ExpenseData } from '@/domain/expense';
+import { database } from '@/firebase';
 
-const ExpensePageContent = lazy(
-  () => import('@/components/containers/expense/ExpensePageContent'),
-);
+type Props = {
+  params: {
+    eventId: string;
+  };
+};
 
-const DashBoard: NextPage = () => {
-  const [event] = useAtom(eventAtom); // TODO: event必須化対応;
+const DashBoard: NextPage<Props> = async ({ params: { eventId } }: Props) => {
+  // TODO: 外部定義
+  const getExpenses = async () => {
+    const docRef = doc(database, eventId, 'expense');
+    try {
+      const document = await getDoc(docRef);
+      const data = document?.data();
+      const expenseList: ExpenseData[] = Object.values(data || {});
+      expenseList.reverse();
+      return expenseList;
+    } catch (error) {
+      throw new Error('Error get document');
+    }
+  };
 
-  return (
-    <Suspense fallback={<div style={{ padding: '24px' }}>Loading...</div>}>
-      <ExpensePageContent eventId={event?.eventId || ''} />
-    </Suspense>
-  );
+  return <ExpensePageContent expenses={await getExpenses()} />;
 };
 
 export default DashBoard;
