+++ src/app/[eventId]/item/page.tsx
@@ -1,26 +1,32 @@
-'use client';
-
-import { useAtom } from 'jotai';
+import { doc, getDoc } from 'firebase/firestore';
 import { NextPage } from 'next';
-import { lazy, Suspense } from 'react';
 
-import { eventAtom } from '@/atoms/eventAtom';
+import ItemPageContent from '@/components/containers/item/ItemPageContent';
+import { database } from '@/firebase';
+import { Data } from '@/hooks/pages/useItemPage';
 
-const ItemPageContent = lazy(
-  () => import('@/components/containers/item/ItemPageContent'),
-);
+type Props = {
+  params: {
+    eventId: string;
+  };
+};
 
-const DashBoard: NextPage = () => {
-  const [event] = useAtom(eventAtom);
+const DashBoard: NextPage<Props> = async ({ params: { eventId } }: Props) => {
+  const getItemList = async () => {
+    const docRef = doc(database, eventId, 'item');
 
-  return (
-    <>
-      {/* TODO: コンポーネント化 */}
-      <Suspense fallback={<div style={{ padding: '24px' }}>Loading...</div>}>
-        <ItemPageContent eventId={event?.eventId || ''} />
-      </Suspense>
-    </>
-  );
+    try {
+      const document = await getDoc(docRef);
+      const data: Data[] = document?.data()?.itemData || [];
+      return data;
+    } catch (error) {
+      throw new Error('Error get document');
+    }
+  };
+
+  const items = await getItemList();
+
+  return <ItemPageContent itemList={items} />;
 };
 
 export default DashBoard;
