+++ src/app/[eventId]/loading.tsx
@@ -1,0 +1,6 @@
+const Loading = () => (
+  <div style={{ padding: '24px' }}>
+    <h1>Loading...</h1>
+  </div>
+);
+export default Loading;
