+++ src/components/containers/memo/MemoPageContent.module.scss
